+++ src/js/defaults.js
@@ -121,8 +121,8 @@ const pluginDefaultOptions = {
     imageOutputQuality: 0.92,
     // Accepts numbers in milliseconds; use this to force intervals-based blobs.
     timeSlice: 0,
-    // Media converter library to use. Legal values are 'ts-ebml' and 'ffmpeg.js'.
-    // Use an empty string '' to disable (default).
+    // Media converter library to use. Legal values are 'ts-ebml', 'ffmpeg.wasm'
+    // and 'ffmpeg.js'. Use an empty string '' to disable (default).
     convertEngine: '',
     // URL for the converter worker.
     convertWorkerURL: '',
@@ -131,7 +131,7 @@ const pluginDefaultOptions = {
     // Enable keyboard hotkeys.
     hotKeys: false,
     // Use this object to specify additional settings for the library used by the
-    // plugin (only used in opus-recorder and vmsg plugins).
+    // plugin (only used in opus-recorder, ffmpeg.js, ffmpeg.wasm and vmsg plugins).
     pluginLibraryOptions: {}
 };