+++ CHANGES.md
@@ -2,8 +2,12 @@
 
 ## 4.2.0 - unreleased
 
-- New opus-media-recorder plugin: provides cross-browser Opus codec support with
-  various audio formats such as Ogg and WebM (#355)
+- New ffmpeg.wasm converter plugin: convert recorded data into other
+  audio/video/image file formats (#522)
+- New opus-media-recorder plugin: provides cross-browser Opus codec
+  support for various audio formats such as Ogg and WebM (#355)
+- ffmpeg.js plugin is deprecated; use new ffmpeg.wasm plugin instead
+- Bump required version for videojs-wavesurfer (3.4.0 or newer)
 
 
 ## 4.1.1 - 2020/11/01